Reimplement get/2 in terms of get/3
This requires modifying the return type of get/3 to include undefined,
but reduces repetition and avoids potential errors in applications using
elli_cookie that call get/2 when Cookies is no_cookies.

Alternatively, we could add a no_cookies clause to get/2, as in get/3,
but this solution is agreeable to me.